[APM] Fixes navigation bugs in APM tutorial/onboarding (#119249)
# Conflicts:
#	x-pack/plugins/apm/public/tutorial/tutorial_fleet_instructions/index.tsx
#	x-pack/plugins/apm/server/routes/fleet/route.ts
