+++ x-pack/plugins/apm/server/tutorial/index.ts
@@ -107,7 +107,11 @@ It allows you to monitor the performance of thousands of applications in real ti
       artifacts,
       customStatusCheckName: 'apm_fleet_server_status_check',
       onPrem: onPremInstructions({ apmConfig, isFleetPluginEnabled }),
-      elasticCloud: createElasticCloudInstructions(cloud),
+      elasticCloud: createElasticCloudInstructions({
+        apmConfig,
+        isFleetPluginEnabled,
+        cloudSetup: cloud,
+      }),
       previewImagePath: '/plugins/apm/assets/apm.png',
       savedObjects,
       savedObjectsInstallMsg: i18n.translate(
